=hco move ObjectRegistry into utils
