+++ akka-http-core/src/main/scala/akka/http/model/HttpCharset.scala
@@ -4,6 +4,7 @@ import scala.language.implicitConversions
 
 import java.nio.charset.Charset
 import akka.http.rendering._
+import akka.http.util.ObjectRegistry
 
 /**
  * A charset range as encountered in `Accept-Charset`. Can either be a single charset, or `*`
