+++ akka-http-core/src/main/scala/akka/http/model/HttpMethod.scala
@@ -1,6 +1,7 @@
 package akka.http.model
 
 import akka.http.rendering.SingletonValueRenderable
+import akka.http.util.ObjectRegistry
 
 /**
  * The method of an HTTP request.
