+++ akka-http-core/src/main/scala/akka/http/model/HttpProtocol.scala
@@ -1,6 +1,7 @@
 package akka.http.model
 
 import akka.http.rendering.SingletonValueRenderable
+import akka.http.util.ObjectRegistry
 
 /** The protocol of an HTTP message */
 case class HttpProtocol private[http] (override val value: String) extends SingletonValueRenderable
