+++ akka-http-core/src/main/scala/akka/http/model/MediaType.scala
@@ -5,6 +5,7 @@ import scala.language.implicitConversions
 import java.util.concurrent.atomic.AtomicReference
 import scala.annotation.tailrec
 import akka.http.rendering._
+import akka.http.util.ObjectRegistry
 
 sealed abstract class MediaRange extends Renderable with WithQValue[MediaRange] {
   def value: String
