+++ akka-http-core/src/main/scala/akka/http/model/StatusCode.scala
@@ -2,6 +2,7 @@ package akka.http.model
 
 import akka.http.rendering.LazyValueBytesRenderable
 import scala.language.implicitConversions
+import akka.http.util.ObjectRegistry
 
 /** The result status code of an HTTP response. */
 sealed abstract class StatusCode extends LazyValueBytesRenderable {
